auth-service/requirements.txt
Flask==2.3.3
PyJWT==2.8.0
bcrypt==4.0.1
redis==4.6.0
cryptography==41.0.4
requests==2.31.0
python-dotenv==1.0.0